added portrait orientation and receiving data from and filling in fields

+++ app/src/main/java/com/example/myfirstapp/DataView.kt
@@ -19,7 +19,7 @@ class DataView @JvmOverloads constructor(
       binding.tvHeadingData.text = heading
     }
 
-    fun setCountLongData(count: Long) {
+    fun setCountLongData(count: Int) {
       binding.tvCountData.text = count.toString()
     }
